+++ castle/cms/pwexpiry/subscriber.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+from AccessControl import Unauthorized
+from plone import api
+from plone.registry.interfaces import IRegistry
+from zope.component import queryUtility
+
+
+
+def ValidPasswordEntered(user, event):
+    pass