+++ HISTORY.rst
@@ -2,8 +2,14 @@
 Changelog
 =========
 
-2.0.46 (unreleased)
--------------------
+2.1 (unreleased)
+----------------
+
+- Add password expiration option with whitelist
+  [CorySanin]
+
+- Fixed a bug with the tab order on the login screen
+  [CorySanin]
 
 - The button for creating a custom content type now defaults to cloning an existing one
   [CorySanin]